DO NOT CAST TO FLOAT (as it is subject to locale settings and you could end up with a comma)

+++ code/fields/AccountingField.php
@@ -118,6 +118,6 @@ class AccountingField extends TextField
         $removedThousendSeparator = preg_replace('/(\.|,)(?=[0-9]{3,}$)/', '',
             $stringWithCommaOrDot);
 
-        return (float) str_replace(',', '.', $removedThousendSeparator);
+        return str_replace(',', '.', $removedThousendSeparator);
     }
 }
